+++ app/packages/frontend/src/components/Button/Button.tsx
@@ -1,0 +1,10 @@
+import React from "react";
+import "./style.css";
+
+export const StartButton = (): JSX.Element => {
+  return (
+    <div className="start-button">
+      <div className="text-wrapper">Begin</div>
+    </div>
+  );
+};
