+++ app/packages/frontend/src/pages/selection-menu.module.css
@@ -1,0 +1,141 @@
+.selection-menu {
+    background-color: #ffffff;
+    display: flex;
+    flex-direction: row;
+    justify-content: center;
+    width: 100%;
+  }
+  
+  .selection-menu .overlap-group-wrapper {
+    background-color: #ffffff;
+    height: 480px;
+    width: 800px;
+  }
+  
+  .selection-menu .overlap-group {
+    border-radius: 25px;
+    height: 480px;
+    position: relative;
+  }
+  
+  .selection-menu .background-logo {
+    -webkit-backdrop-filter: blur(4px) brightness(100%);
+    backdrop-filter: blur(4px) brightness(100%);
+    background-color: #333131;
+    border-radius: 25px;
+    height: 480px;
+    left: 0;
+    overflow: hidden;
+    position: absolute;
+    top: 0;
+    width: 800px;
+  }
+  
+  .selection-menu .HAL-logo {
+    height: 480px;
+    left: 111px;
+    object-fit: cover;
+    position: absolute;
+    top: 0;
+    width: 577px;
+  }
+  
+  .selection-menu .button {
+    all: unset;
+    background-color: #013143;
+    border: 1px solid;
+    border-color: #ffffff1a;
+    border-radius: 25px;
+    box-sizing: border-box;
+    height: 80px;
+    left: 216px;
+    overflow: hidden;
+    position: absolute;
+    top: 340px;
+    width: 397px;
+  }
+  
+  .selection-menu .text-wrapper {
+    color: #a49b95;
+    font-family: "JetBrains Mono-ExtraBold", Helvetica;
+    font-size: 32px;
+    font-weight: 800;
+    left: 40px;
+    letter-spacing: 0;
+    line-height: 110.5px;
+    position: absolute;
+    text-align: center;
+    top: -18px;
+    white-space: nowrap;
+    width: 318px;
+  }
+  
+  .selection-menu .div-wrapper {
+    all: unset;
+    background-color: #013143;
+    border: 1px solid;
+    border-color: #ffffff1a;
+    border-radius: 25px;
+    box-sizing: border-box;
+    height: 80px;
+    left: 216px;
+    overflow: hidden;
+    position: absolute;
+    top: 200px;
+    width: 397px;
+  }
+  
+  .selection-menu .div {
+    color: #a49b95;
+    font-family: "JetBrains Mono-ExtraBold", Helvetica;
+    font-size: 32px;
+    font-weight: 800;
+    left: 40px;
+    letter-spacing: 0;
+    line-height: 112px;
+    position: absolute;
+    text-align: center;
+    top: -18px;
+    white-space: nowrap;
+    width: 318px;
+  }
+  
+  .selection-menu .button-2 {
+    all: unset;
+    background-color: #013143;
+    border: 1px solid;
+    border-color: #ffffff1a;
+    border-radius: 25px;
+    box-sizing: border-box;
+    height: 80px;
+    left: 216px;
+    overflow: hidden;
+    position: absolute;
+    top: 60px;
+    width: 397px;
+  }
+  
+  .selection-menu .text-wrapper-2 {
+    color: #a49b95;
+    font-family: "JetBrains Mono-ExtraBold", Helvetica;
+    font-size: 32px;
+    font-weight: 800;
+    left: 40px;
+    letter-spacing: 0;
+    line-height: 114px;
+    position: absolute;
+    text-align: center;
+    top: -18px;
+    white-space: nowrap;
+    width: 318px;
+  }
+  
+  .selection-menu .img {
+    height: 94px;
+    left: 706px;
+    object-fit: cover;
+    position: absolute;
+    top: 386px;
+    width: 94px;
+  }
+
