+++ app/packages/frontend/src/pages/spotify-menu.module.css
@@ -1,0 +1,51 @@
+.spotify-menu {
+    background-color: #ffffff;
+    display: flex;
+    flex-direction: row;
+    justify-content: center;
+    width: 100%;
+  }
+  
+  .spotify-menu .overlap-group-wrapper {
+    background-color: #ffffff;
+    height: 480px;
+    width: 800px;
+  }
+  
+  .spotify-menu .overlap-group {
+    border-radius: 25px;
+    height: 480px;
+    position: relative;
+  }
+  
+  .spotify-menu .background-logo {
+    -webkit-backdrop-filter: blur(4px) brightness(100%);
+    backdrop-filter: blur(4px) brightness(100%);
+    background-color: #333131;
+    border-radius: 25px;
+    height: 480px;
+    left: 0;
+    overflow: hidden;
+    position: absolute;
+    top: 0;
+    width: 800px;
+  }
+  
+  .spotify-menu .HAL-logo {
+    height: 480px;
+    left: 111px;
+    object-fit: cover;
+    position: absolute;
+    top: 0;
+    width: 577px;
+  }
+  
+  .spotify-menu .img {
+    height: 94px;
+    left: 706px;
+    object-fit: cover;
+    position: absolute;
+    top: 386px;
+    width: 94px;
+  }
+
