+++ package.json
@@ -3,6 +3,8 @@
     "@types/react": "^18.2.42",
     "@types/react-dom": "^18.2.17",
     "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.20.1",
     "react-webcam": "^7.2.0"
   }
 }
